Fix saving config file.
This is yet another DLang struct/class constructor workaround. Classes can have a default destructor but the destructor
is not guaranteed to be called and also a default constructor. Structs on the other hand can't have a default contructor
and the destructor is always called on when object goes out of scope.